like and del card

+++ src/utils/api.js
@@ -69,24 +69,24 @@ class Api {
     })
   }
 
-  likeCard({ cardId }) {
-    return fetch(this._baseUrl + `/cards/${cardId}/likes`, {
-      method: 'PUT',
-      headers: this._headers
-    })
-    .then((res) => {
-      return this._getResponseData(res);
-    })
-  }
-
-  dislikeCard({ cardId }) {
-    return fetch(this._baseUrl + `/cards/${cardId}/likes`, {
-      method: 'DELETE',
-      headers: this._headers
-    })
-    .then((res) => {
-      return this._getResponseData(res);
-    })
+  changeLikeCardStatus({ cardId, isLiked }) {
+    if (isLiked) {
+      return fetch(this._baseUrl + `/cards/${cardId}/likes`, {
+        method: 'PUT',
+        headers: this._headers
+      })
+      .then((res) => {
+        return this._getResponseData(res);
+      })
+    } else {
+      return fetch(this._baseUrl + `/cards/${cardId}/likes`, {
+        method: 'DELETE',
+        headers: this._headers
+      })
+      .then((res) => {
+        return this._getResponseData(res);
+      })
+    }
   }
 
   changeAvatar({ avatarPopupInputValue }) {
